refactor(utils): move env code to env_utils
