+++ tests/unit/core/test_utils.py
@@ -1,7 +1,7 @@
 import os
 import sys
-from typing import AnyStr, Dict, Optional, Tuple
-from unittest.mock import Mock, patch
+from typing import AnyStr, Tuple
+from unittest.mock import patch
 
 import click
 import pytest
@@ -10,8 +10,6 @@ from ggshield.core.client import create_client_from_config
 from ggshield.core.config import Config
 from ggshield.core.errors import APIKeyCheckError, UnexpectedError, UnknownInstanceError
 from ggshield.core.scan.scan_context import parse_os_release
-from ggshield.core.utils import load_dot_env
-from ggshield.verticals.secret.repo import cd
 
 
 def test_retrieve_client_invalid_api_url():
@@ -69,85 +67,6 @@ def test_retrieve_client_unknown_custom_dashboard_url(isolated_fs):
             config = Config()
             config.cmdline_instance_name = "https://example.com"
             create_client_from_config(config)
-
-
-@patch("ggshield.core.utils.load_dotenv")
-@pytest.mark.parametrize(
-    ["cwd", "env", "expected_dotenv_path"],
-    [
-        ("sub", {}, None),
-        (".", {}, ".env"),
-        (".", {"GITGUARDIAN_DONT_LOAD_ENV": "1"}, None),
-        (".", {"GITGUARDIAN_DOTENV_PATH": ".custom-env"}, ".custom-env"),
-        (".", {"GITGUARDIAN_DOTENV_PATH": ".does-not-exist"}, None),
-    ],
-)
-def test_load_dot_env(
-    load_dotenv_mock: Mock,
-    monkeypatch,
-    tmp_path,
-    cwd: str,
-    env: Dict[str, str],
-    expected_dotenv_path: Optional[str],
-):
-    """
-    GIVEN a file hierarchy like this:
-    /sub/
-    /.env
-    /.custom-env
-    AND environment variables set according to `env`
-    AND the current working directory being `/sub/`
-    WHEN load_dot_env() is called
-    THEN the appropriate environment file is loaded
-    """
-    (tmp_path / "sub").mkdir()
-    (tmp_path / ".env").touch()
-    (tmp_path / ".custom-env").touch()
-
-    if env:
-        for key, value in env.items():
-            monkeypatch.setenv(key, value)
-
-    with cd(str(tmp_path / cwd)):
-        load_dot_env()
-
-        if expected_dotenv_path:
-            expected_dotenv_path = tmp_path / expected_dotenv_path
-            load_dotenv_mock.assert_called_once_with(
-                str(expected_dotenv_path), override=True
-            )
-        else:
-            load_dotenv_mock.assert_not_called()
-
-
-@patch("ggshield.core.utils.get_git_root")
-@patch("ggshield.core.utils.is_git_dir")
-@patch("ggshield.core.utils.load_dotenv")
-def test_load_dot_env_loads_git_root_env(
-    load_dotenv_mock: Mock, is_git_dir_mock, get_git_root_mock, tmp_path
-):
-    """
-    GIVEN a git repository checkout with this file hierarchy:
-    /sub1/sub2
-    /sub1/.env
-    /.env
-    AND the current working directory being `/sub1/sub2`
-    WHEN load_dot_env() is called
-    THEN the .env file at the root of the git repository is loaded
-    """
-    is_git_dir_mock.return_value = True
-    get_git_root_mock.return_value = str(tmp_path)
-
-    sub1_sub2_dir = tmp_path / "sub1" / "sub2"
-    git_root_dotenv = tmp_path / ".env"
-
-    sub1_sub2_dir.mkdir(parents=True)
-    (tmp_path / "sub1" / ".env").touch()
-    git_root_dotenv.touch()
-
-    with cd(str(sub1_sub2_dir)):
-        load_dot_env()
-        load_dotenv_mock.assert_called_once_with(str(git_root_dotenv), override=True)
 
 
 @pytest.mark.skipif(
